Move parameter validation to middleware
