+++ connexion/operations/openapi.py
@@ -35,7 +35,6 @@ class OpenAPIOperation(AbstractOperation):
         validate_responses=False,
         strict_validation=False,
         randomize_endpoint=None,
-        validator_map=None,
         pythonic_params=False,
         uri_parser_class=None,
     ):
@@ -72,8 +71,6 @@ class OpenAPIOperation(AbstractOperation):
         :type strict_validation: bool
         :param randomize_endpoint: number of random characters to append to operation name
         :type randomize_endpoint: integer
-        :param validator_map: Custom validators for the types "parameter", "body" and "response".
-        :type validator_map: dict
         :param pythonic_params: When True CamelCase parameters are converted to snake_case and an underscore is appended
             to any shadowed built-ins
         :type pythonic_params: bool
@@ -97,7 +94,6 @@ class OpenAPIOperation(AbstractOperation):
             validate_responses=validate_responses,
             strict_validation=strict_validation,
             randomize_endpoint=randomize_endpoint,
-            validator_map=validator_map,
             pythonic_params=pythonic_params,
             uri_parser_class=uri_parser_class,
         )
